Create agent.step() that captures env.step() and memory.append()
